Split the e2e test from the other tests
