+++ package.json
@@ -8,6 +8,8 @@
     "lint:md": "markdownlint **/*.md --ignore node_modules",
     "lint:ts": "eslint --ext .ts,.js .",
     "lint:yaml": "yamllint --ignore=node_modules/**/*.yaml **/*.yaml",
+    "test:e2e": "mocha \"tests/e2e/**/*.test.ts\"",
+    "test:src": "nyc --extension .ts mocha \"tests/src/**/*.test.ts\"",
     "typecheck": "tsc --noEmit -P tests/tsconfig.json "
   },
   "dependencies": {
