+++ tests/src/utils/gameStateUtil.test.ts
@@ -1,9 +1,9 @@
 import { expect } from 'chai';
 import sinon, { SinonSpy } from 'sinon';
-import { databaseUtil } from '../../src/utils/databaseUtil';
-import { gameStateUtil } from '../../src/utils/gameStateUtil';
-import { GameState } from '../../src/utils/types';
-import { generateGameState } from '../testHelpers/generateGameState';
+import { databaseUtil } from '../../../src/utils/databaseUtil';
+import { gameStateUtil } from '../../../src/utils/gameStateUtil';
+import { GameState } from '../../../src/utils/types';
+import { generateGameState } from '../../testHelpers/generateGameState';
 
 describe('gameStateUtil', () => {
   describe('load', () => {
